third_party/blink/web_tests/fast/loader/scroll-restore-should-happen-during-load-expected.txt

Test ensures that frame scroll position is only restored during the navigation process but does not occur after navigation is complete for same document navigations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement.scrollLeft is 0
PASS document.scrollingElement.scrollTop is 0
Verify that scroll position restored correctly during navigation.
PASS document.scrollingElement.scrollLeft is 100
PASS document.scrollingElement.scrollTop is 100
PASS document.scrollingElement.scrollLeft is 0
PASS document.scrollingElement.scrollTop is 0
Verify that scroll position is not restored again after navigation is complete.
PASS document.scrollingElement.scrollLeft is 0
PASS document.scrollingElement.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
